requirements.txt
amadeus==12.0.0
python-dotenv==1.1.1
langchain==0.3.26
langchain-core==0.3.68
langchain-google-genai==2.1.7
google-ai-generativelanguage==0.6.18

